2. Are there hooks to allow the title to be rendered differently? could an interface be put in if not? the default styling is nice, but doesn't go with my app (see screenshot).  I'd like to customise the rendering.
   3. is there a way to allow a drop down menu to replace the (up to)
5 buttons at the right of each title bar? e.g. I find the current look
a bit cluttered, and I'd love to have a single icon which brought up a
popup list of choices.  Possibly the hooks for (2) could cover this as
well.  e.g. an interface for rendering the titles (including the
buttons), which has a default implementation which renders in the
current way.

git-svn-id: https://svn.code.sf.net/p/mydoggy/code/mydoggy/trunk@673 2a62bbde-051e-0410-bf3e-cfb6b34e140d