fix: do not update the select tab from TabbedContentPane
When tabs are spread on multiple rows and when a tab is clicked, some
tabs dimensions will change (adding/removing the detach/close buttons)
therefore the position of the clicked tab might change.
In that case updating the selection from TabbedContentPane interfers
with Swing's native calculation of the clicked tab, because the X/Y
coordinates now correspond to another tab.
In the end the user clicked a tab and another was selected.